requirements.txt
numpy>=1.24
pandas>=2.0
scipy>=1.10
scikit-learn>=1.3
matplotlib>=3.7
seaborn>=0.12
pyyaml>=6.0
xgboost>=2.0
lifelines>=0.28
shap>=0.43
torch>=2.0
